requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-digitalocean==1.17.0
paramiko==3.4.0
pydantic==2.5.3
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
psycopg2-binary==2.9.9
cryptography==42.0.0
httpx==0.26.0
